+++ src/records/SeederUserRecord.php
@@ -4,13 +4,21 @@ namespace anubarak\seeder\records;
 
 use craft\db\ActiveRecord;
 
+/**
+ * Class SeederUserRecord
+ *
+ * @package anubarak\seeder\records
+ * @since   15.07.2024
+ * @author  by Robin Schambach
+ * @property string $userUid
+ */
 class SeederUserRecord extends ActiveRecord
 {
 
 	// Props
 	// =========================================================================
 
-	public static $tableName = '{{%seeder_user}}';
+	public static string $tableName = '{{%seeder_user}}';
 
 	/**
 	 * @inheritdoc